Car owns his position

+++ src/Vector3.cpp
@@ -6,7 +6,7 @@ Vector3::Vector3() {
 	_z=0.0;
 }
 
-Vector3::Vector3(double x, double y, double z) {
+Vector3::Vector3(float x, float y, float z) {
 	_x=x;
 	_y=y;
 	_z=z;
@@ -14,31 +14,31 @@ Vector3::Vector3(double x, double y, double z) {
 
 Vector3::~Vector3() {}
 
-double Vector3::getX() const {
+float Vector3::getX() const {
 		return _x;
 }
 
-double Vector3::getY() const {
+float Vector3::getY() const {
 		return _y;
 }
 
-double Vector3::getZ() const {
+float Vector3::getZ() const {
 		return _z;
 }
 
-void Vector3::setX(double x) {
+void Vector3::setX(float x) {
 	_x = x;
 }
 
-void Vector3::setY(double y) {
+void Vector3::setY(float y) {
 	_y = y;
 }
 
-void Vector3::setZ(double z) {
+void Vector3::setZ(float z) {
 	_z = z;
 }
 
-void Vector3::set(double x, double y, double z) {
+void Vector3::set(float x, float y, float z) {
 	_x=x;
 	_y=y;
 	_z=z;
